% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_ranges.R
\name{group_overlapping_ranges}
\alias{group_overlapping_ranges}
\title{Group overlapping ranges into episodes}
\usage{
group_overlapping_ranges(group, range_start, range_stop)
}
\arguments{
\item{group}{A vector specifying the grouping variable for each range. Ranges with the same group value
can potentially belong to the same episode, while ranges with different groups cannot overlap
or share an episode. This vector must be of the same length as `range_start` and `range_stop`.
Acceptable types: character, factor, or numeric.}

\item{range_start}{A vector of start dates or times for each range. This vector must be the same length as `group`
and `range_stop`, and should be sortable (e.g., Date, POSIXct, numeric).}

\item{range_stop}{A vector of end dates or times for each range, with each value corresponding to a `range_start`
value. This vector must be the same length as `group` and `range_start` and should have the same
data type as `range_start`. Each `range_stop` date should be greater than or equal to its corresponding
`range_start`.}
}
\value{
A vector of episode numbers, where each unique integer identifies an episode of overlapping ranges
        within each group. The length of the returned vector matches the length of the input parameters.
}
\description{
This function assigns episode numbers to ranges that overlap in time, grouped by a specified variable.
Each unique episode number identifies a set of overlapping ranges within the same group.
}
